+++ bootstrap/src/test/java/bootstrap/BundleTest.java
@@ -1,3 +1,5 @@
+package bootstrap;
+
 import org.junit.Assert;
 import org.junit.Test;
