+++ handlebars/src/test/java/handlebars/BundleTest.java
@@ -1,7 +1,10 @@
+package handlebars;
+
 import org.junit.Assert;
 import org.junit.Test;
 
 import java.io.File;
+import java.util.Arrays;
 
 /**
  * @author <a href="mailto:tom.fennelly@gmail.com">tom.fennelly@gmail.com</a>
@@ -10,7 +13,6 @@ public class BundleTest {
     
     @Test
     public void test_core_assest_assembly() {
-        Assert.assertTrue(new File("./target/assets-wrapper/assets/ace-editor/jsmodules/ace-editor-119.js").exists());
-        Assert.assertTrue(new File("./target/assets-wrapper/assets/ace-editor/jsmodules/ace-editor-122.js").exists());
+        Assert.assertTrue(new File("./target/assets-wrapper/assets/handlebars/jsmodules/handlebars3.js").exists());
     }
 }
