+++ jquery-detached/src/test/java/jquery/BundleTest.java
@@ -1,3 +1,5 @@
+package jquery;
+
 import org.junit.Assert;
 import org.junit.Test;
