+++ momentjs/src/test/java/momentjs/BundleTest.java
@@ -1,3 +1,5 @@
+package momentjs;
+
 import org.junit.Assert;
 import org.junit.Test;
