+++ numeraljs/src/test/java/numeraljs/BundleTest.java
@@ -1,3 +1,5 @@
+package numeraljs;
+
 import org.junit.Assert;
 import org.junit.Test;
